% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antider_logis.R
\name{antider_logis}
\alias{antider_logis}
\title{Anti derivative of the logistic function}
\usage{
antider_logis(x, max, infl, slope)
}
\arguments{
\item{x}{Value to calculate.}

\item{max}{Asymptote of the logistic function.}

\item{infl}{Inflection point of the logistic function.}

\item{slope}{Slope of the logistic function.}
}
\value{
Numeric value
}
\description{
Anti derivative of the logistic function
}
