% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsinsqrt.R
\name{arcsinsqrt}
\alias{arcsinsqrt}
\title{Arc sin transformation}
\usage{
arcsinsqrt(x)
}
\arguments{
\item{x}{Value to transform.}
}
\value{
The arc sin square root transformation of x.
}
\description{
Performs the arc sin square root transformation.
}
