% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_squarelike.R
\name{area_squarelike}
\alias{area_squarelike}
\title{Area under the squarelike function}
\usage{
area_squarelike(
  lower_bound,
  upper_bound,
  l_inflection,
  l_slope,
  r_inflection,
  r_slope,
  max
)
}
\arguments{
\item{lower_bound}{Value for the lower bound of area.}

\item{upper_bound}{Value for the upper bound of area.}

\item{l_inflection}{The left inflection point.}

\item{l_slope}{The slope of the left side of the function.}

\item{r_inflection}{The right inflection point.}

\item{r_slope}{The slope of the right side of the function.}

\item{max}{The maximum value for the function.}
}
\value{
numeric value for the area under the squarelike function
}
\description{
Area under the squarelike function
}
