% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_value.R
\name{clip_value}
\alias{clip_value}
\title{Clip value to between bounds}
\usage{
clip_value(x, l, u)
}
\arguments{
\item{x}{Value to clip.}

\item{l}{Lower bound for clipping.}

\item{u}{Upper bound for clipping.}
}
\value{
Clipped value between l and u.
}
\description{
Clip value to between bounds
}
