% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_logs.R
\name{combine_logs}
\alias{combine_logs}
\title{Combine log files into a single file and delete old files}
\usage{
combine_logs(log_directory)
}
\arguments{
\item{log_directory}{directory where to find the log files}
}
\value{
Nothing
}
\description{
Combine log files into a single file and delete old files
}
