% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_area.R
\docType{class}
\name{constant_area-class}
\alias{constant_area-class}
\title{Class to hold the initial area as a constant}
\description{
Class to hold the initial area as a constant
}
\section{Slots}{

\describe{
\item{\code{lower_bound}}{Value for the lower bound of integration.}

\item{\code{upper_bound}}{Value for the upper bound of integration.}

\item{\code{initial_area}}{Value of area that will be kept constant}
}}

