% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_area.R
\name{constant_area}
\alias{constant_area}
\title{Constructor for constant_area}
\usage{
constant_area(lower_bound, upper_bound, func, ...)
}
\arguments{
\item{lower_bound}{Value for the lower bound of integration.}

\item{upper_bound}{Value for the upper bound of integration.}

\item{func}{Function to use for calculation}

\item{...}{Arguments for the function}
}
\value{
a constant_area object
}
\description{
Constructor for constant_area
}
