% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_cor.R
\name{custom_cor}
\alias{custom_cor}
\title{Bootstrap correlation}
\usage{
custom_cor(d, i)
}
\arguments{
\item{d}{The dataset object to use.}

\item{i}{The row in the dataset to calculate the correlation.}
}
\value{
The correlation for the selected row.
}
\description{
Calculate correlations from a dataset with two columns. Used for bootstrap.
}
