% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{fecundity}
\alias{fecundity}
\title{Fly fecundity}
\usage{
fecundity(x, mean, sd, skew)
}
\arguments{
\item{x}{Value of x to calculate the fecundity.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{skew}{The skew parameter for a skewed normal distribution.}
}
\value{
Returns the value of fecundity for x.
}
\description{
Calculate the fecundity of a fly by calculating a skewed normal pdf.
}
