% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{find_density_peaks}
\alias{find_density_peaks}
\title{Finds the density peaks in the provided vector}
\usage{
find_density_peaks(numbers, npeaks = NULL, decreasing = TRUE)
}
\arguments{
\item{numbers}{The vector for which to find the peaks.}

\item{npeaks}{How many peaks to output.}

\item{decreasing}{If the order should be decreasing or not.}
}
\value{
The density npeaks in the provided vector
}
\description{
Finds the density peaks in the provided vector
}
