% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\docType{class}
\name{fly-class}
\alias{fly-class}
\title{Fly class}
\description{
Class to keep individual values together inside a single "fly".
Will be created with the specific values for each genomic trait and the
phenotype will be calculated with internal functions
}
\section{Slots}{

\describe{
\item{\code{replicate}}{The simulation replicate the fly was in.}

\item{\code{population}}{Population in which the fly was in.}

\item{\code{min_pe}}{Minimum value for the prediction interval.}

\item{\code{max_pe}}{Maximum value for the prediction interval.}

\item{\code{mid_pe}}{Midpoint value for the prediction interval.}

\item{\code{total_pe}}{Total length value for the prediction interval.}

\item{\code{precision}}{Precision for all calculations.}

\item{\code{x_pe}}{The range of used temperatures.}

\item{\code{pe_genes}}{Vector of 25 values that define the PE reaction norm.}

\item{\code{gw_genes}}{Vector of 5 vales that define the GW reaction norm.}

\item{\code{ow_genes}}{Vector of 5 vales that define the OW reaction norm.}

\item{\code{pw_genes}}{Vector of 5 vales that define the PW reaction norm.}

\item{\code{mean_surv}}{Genetic value for the survival midpoint.}

\item{\code{PE}}{PE reaction norm.}

\item{\code{GW}}{GW reaction norm.}

\item{\code{OW}}{OW reaction norm.}

\item{\code{PW}}{PW reaction norm.}
}}

