% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{fly}
\alias{fly}
\title{Constructor method for fly}
\usage{
fly(
  replicate,
  population,
  min_pe,
  max_pe,
  precision,
  x_pe,
  pe_genes,
  gw_genes,
  ow_genes,
  pw_genes,
  mean_surv,
  inverted_rnorm = FALSE
)
}
\arguments{
\item{replicate}{The simulation replicate the fly was in.}

\item{population}{Population in which the fly was in.}

\item{min_pe}{min_PE Minimum value for the prediction interval.}

\item{max_pe}{Maximum value for the prediction interval.}

\item{precision}{Precision for all calculations.}

\item{x_pe}{The range of used temperatures.}

\item{pe_genes}{Vector of 25 values that define the PE reaction norm.}

\item{gw_genes}{Vector of 5 vales that define the GW reaction norm.}

\item{ow_genes}{Vector of 5 vales that define the OW reaction norm.}

\item{pw_genes}{Vector of 5 vales that define the PW reaction norm.}

\item{mean_surv}{Genetic value for the survival midpoint.}

\item{inverted_rnorm}{Logical value for if the simulation was done with
inverted reaction norms (intercept on the right side of the x-axis)}
}
\description{
Constructor method for fly
}
