% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{generate_reps}
\alias{generate_reps}
\title{Digit expansion}
\usage{
generate_reps(rep_group)
}
\arguments{
\item{rep_group}{Vector with tens to expand.}
}
\value{
A vector with the expanded group of numbers.
}
\description{
Expands a vector of tens to have all digits (19 -> 190, 191,192,...).
}
