% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean}
\usage{
geometric_mean(x, na_rm = TRUE)
}
\arguments{
\item{x}{The vector for which to calculate the geometric mean.}

\item{na_rm}{If NAs should be removed from the calculation. Default is TRUE.}
}
\value{
The geometric mean of the vector.
}
\description{
Calculates the geometric mean of a vector of values.
}
