% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{get_fecundity}
\alias{get_fecundity}
\alias{get_fecundity,fly-method}
\title{fly method to retrieve fecundity with day}
\usage{
get_fecundity(object, day)

\S4method{get_fecundity}{fly}(object, day)
}
\arguments{
\item{object}{An object}

\item{day}{Day for which to retrieve fecundity.}
}
\value{
Fecundity at the provided day of year cycle.

Fecundity at the provided day of year cycle.
}
\description{
fly method to retrieve fecundity with day

fly method to retrieve fecundity with day
}
