% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{get_prediction}
\alias{get_prediction}
\title{fly method to retrieve prediction with two temperatures}
\usage{
get_prediction(object, first_temp, second_temp)
}
\arguments{
\item{object}{An object}

\item{first_temp}{Value of temperature for first fly experience.}

\item{second_temp}{Value of temperature forsecond fly experience.}
}
\description{
fly method to retrieve prediction with two temperatures
}
