% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_area.R
\name{get_scale}
\alias{get_scale}
\alias{get_scale,constant_area-method}
\title{Method to compare areas to initial area}
\usage{
get_scale(func, ...)

\S4method{get_scale}{constant_area}(object, func, ...)
}
\arguments{
\item{func}{Function to use for calculation}

\item{...}{Arguments for the function}

\item{object}{The constant_area object to use.}
}
\value{
Numeric scale of the comparison betwee initial and provided areas.
}
\description{
Method to compare areas to initial area
}
