% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{get_weight}
\alias{get_weight}
\alias{get_weight,fly-method}
\title{fly method to retrieve information weights}
\usage{
get_weight(
  object,
  temperature,
  weight,
  offspring_error = NULL,
  presampled_error = NULL
)

\S4method{get_weight}{fly}(
  object,
  temperature,
  weight,
  offspring_error = NULL,
  presampled_error = NULL
)
}
\arguments{
\item{object}{An object}

\item{temperature}{Value of temperature.}

\item{weight}{Which weight to retrive. "GW", "OW", or "PW".}

\item{offspring_error}{Amount of error that offspring have to measure
temperature. Varies between replicates.}

\item{presampled_error}{Pre-sampled error in cases where randomness will
be re-used}
}
\value{
The information weight requested for the provided temperature.
}
\description{
fly method to retrieve information weights
}
