% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse of the \%in\% operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{A vector of values to look for in \code{table}.}

\item{table}{A vector where \code{x} values might be found.}
}
\value{
A logical vector indicating if a match was not located for each
element of \code{x} in \code{table}.
}
\description{
Inverse of the \%in\% operator
}
\examples{
# Check if values are in a vector
c(1, 2, 3) \%notin\% c(1, 2, 5)  # Returns: FALSE FALSE TRUE
}
