% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_value.R
\name{invert_value}
\alias{invert_value}
\title{Inverting a number around a specific value}
\usage{
invert_value(x, invert_axis)
}
\arguments{
\item{x}{The value to be inverted.}

\item{invert_axis}{The axis to which x will be inverted around.}
}
\value{
A numeric value inverted around invert_axis.
}
\description{
Function to invert numeric values around a specific value.
}
