% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_norm_pdf.R
\name{normal_cdf}
\alias{normal_cdf}
\title{Normal distribution cdf}
\usage{
normal_cdf(x)
}
\arguments{
\item{x}{Numeric value for which to calculate the normal cdf.}
}
\value{
The normal cdf for x.
}
\description{
Calculate the normal distribution cumulative density function at a specific
x value.
}
