% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_norm_pdf.R
\name{normal_pdf}
\alias{normal_pdf}
\title{Normal distribution pdf}
\usage{
normal_pdf(x)
}
\arguments{
\item{x}{Numeric value for which to calculate the normal pdf.}
}
\value{
The normal pdf for x.
}
\description{
Calculate the normal distribution probability density function at a specific
x value.
}
