% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive.R
\name{positive}
\alias{positive}
\title{Transform values to positive or zero}
\usage{
positive(x)
}
\arguments{
\item{x}{numeric value to transform}
}
\value{
transformed numeric value
}
\description{
Transform values to positive or zero
}
