% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{round_any}
\alias{round_any}
\title{Round any}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{Value orr vector to round.}

\item{accuracy}{Value with which to round.}

\item{f}{Function to use. Can be floor, ceiling, or round. Default is round.}
}
\value{
The transformed value.
}
\description{
Rounds, floors, or ceilings with provided accuracy. Copied from plyr.
}
