% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vec.R
\name{sample_vec}
\alias{sample_vec}
\title{Returns a random sample from a vector}
\usage{
sample_vec(x, ...)
}
\arguments{
\item{x}{Vector to sample from}

\item{...}{Arguments for the sample function}
}
\value{
Returns sampled values from the supplied vector
}
\description{
Returns a random sample from a vector
}
