% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{sine_wave}
\alias{sine_wave}
\title{Sine wave}
\usage{
sine_wave(base_env, amplitude, day, period)
}
\arguments{
\item{base_env}{The base value around which the sine wave will change.}

\item{amplitude}{The amplitude of the sine wave.}

\item{day}{The location in the sine wave to be calculated.}

\item{period}{The period of the sine wave.}
}
\value{
Returns the value of the sine wave on the specified day.
}
\description{
Calculates the sine value at point x. Incorporates a base value, amplitude,
and period.
}
