% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_norm_pdf.R
\name{skew_normal_pdf}
\alias{skew_normal_pdf}
\title{Skew normal distribution pdf}
\usage{
skew_normal_pdf(x, mean, sd, skew)
}
\arguments{
\item{x}{The x value.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{skew}{The skew parameter for a skewed normal distribution.}
}
\value{
The normal pdf for x.
}
\description{
Calculate the skew normal distribution probability density function at a
specific x value.
}
