% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{skewnormal_cdf}
\alias{skewnormal_cdf}
\title{Skew normal distribution cdf}
\usage{
skewnormal_cdf(mean, sd, x, skew)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{x}{The x value.}

\item{skew}{The skew parameter for a skewed normal distribution.}
}
\value{
The normal cdf for x.
}
\description{
Calculate the normal distribution cumulative density function at a specific x value.
}
\examples{
normal_cdf(0, 1, 0.5, -5)
normal_cdf(10, 3, 9, 2)
}
