% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droso_utils.R
\name{skewnormal_pdf}
\alias{skewnormal_pdf}
\title{Skew normal distribution pdf}
\usage{
skewnormal_pdf(mean, sd, x)
}
\arguments{
\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{x}{The x value.}
}
\value{
The normal pdf for x.
}
\description{
Calculate the normal distribution probability density function at a specific x value.
}
