% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squarelike.R
\name{squarelike}
\alias{squarelike}
\title{Square-like function}
\usage{
squarelike(x, l_inflection, l_slope, r_inflection, r_slope, max)
}
\arguments{
\item{x}{Point or vector to calculate the function.}

\item{l_inflection}{The left inflection point.}

\item{l_slope}{The slope of the left side of the function.}

\item{r_inflection}{The right inflection point.}

\item{r_slope}{The slope of the right side of the function.}

\item{max}{The maximum value for the function.}
}
\value{
The y value for x.
}
\description{
Calculates the output using a square-like shaped curve.
}
