% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_mean.R
\name{std_mean}
\alias{std_mean}
\title{Standard error of the mean}
\usage{
std_mean(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The standard error of the mean from the vector.
}
\description{
Calculates the standard error of the mean.
}
