% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{Fly survival}
\usage{
survival(temperature, experience)
}
\arguments{
\item{temperature}{The current temperature for which to test survival}

\item{experience}{The phenotype (midpoint) of the fly to be tested.}
}
\value{
The probability of survival of the fly.
}
\description{
Calculates the survival of a fly by creating a square-shaped function with
two logistic curves and a modified normal distribution pdf.
}
