% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{calculate_phenotype}
\alias{calculate_phenotype}
\alias{calculate_phenotype,fly-method}
\title{fly method to calculate phenotype}
\usage{
calculate_phenotype(object, univar_matrix, bivar_matrix)

\S4method{calculate_phenotype}{fly}(object, univar_matrix, bivar_matrix)
}
\arguments{
\item{object}{An object}

\item{univar_matrix}{1D matrix from rcspline.}

\item{bivar_matrix}{2D matrix from rcspline.}
}
\value{
The same object but modified to have a calculated phenotype.

The same object but modified to have a calculated phenotype.
}
\description{
fly method to calculate phenotype

fly method to calculate phenotype
}
