% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{calculate_phenotype}
\alias{calculate_phenotype}
\title{fly method to calculate phenotype}
\usage{
calculate_phenotype(object, univar_matrix, bivar_matrix, logis_k, logis_x0)
}
\arguments{
\item{object}{An object}

\item{univar_matrix}{1D matrix from rcspline.}

\item{bivar_matrix}{2D matrix from rcspline.}

\item{logis_k}{Value for the k parameter for the inverse-logit function.}

\item{logis_x0}{Value for the x0 parameter for the inverse-logit function.}
}
\description{
fly method to calculate phenotype
}
