% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\docType{class}
\name{fly-class}
\alias{fly-class}
\title{Fly class}
\description{
Class to keep individual values together inside a single "fly".
Will be created with the specific values for each genomic trait and the
phenotype will be calculated with internal functions
}
\section{Slots}{

\describe{
\item{\code{replicate}}{The simulation replicate the fly was in.}

\item{\code{population}}{Population in which the fly was in.}

\item{\code{min_PE}}{Minimum value for the prediction interval.}

\item{\code{max_PE}}{Maximum value for the prediction interval.}

\item{\code{mid_PE}}{Midpoint value for the prediction interval.}

\item{\code{total_PE}}{Total length value for the prediction interval.}

\item{\code{precision}}{Precision for all calculations.}

\item{\code{x_PE}}{The range of used temperatures.}

\item{\code{PE_genes}}{Vector of 25 values that define the PE reaction norm.}

\item{\code{GW_genes}}{Vector of 5 vales that define the GW reaction norm.}

\item{\code{OW_genes}}{Vector of 5 vales that define the OW reaction norm.}

\item{\code{PW_genes}}{Vector of 5 vales that define the PW reaction norm.}

\item{\code{mean_surv}}{Genetic value for the survival midpoint.}

\item{\code{fecundity_genes}}{Vector of 3 values that define the mean,
sd and skew for fecundity.}

\item{\code{PE}}{PE reaction norm.}

\item{\code{GW}}{GW reaction norm.}

\item{\code{OW}}{OW reaction norm.}

\item{\code{PW}}{PW reaction norm.}

\item{\code{egg_laying}}{Reaction for egg laying.}
}}

