% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{fly}
\alias{fly}
\title{Constructor method for fly}
\usage{
fly(
  replicate,
  population,
  min_PE,
  max_PE,
  precision,
  x_PE,
  PE_genes,
  GW_genes,
  OW_genes,
  PW_genes,
  mean_surv,
  fecundity_genes
)
}
\arguments{
\item{replicate}{The simulation replicate the fly was in.}

\item{population}{Population in which the fly was in.}

\item{max_PE}{Maximum value for the prediction interval.}

\item{precision}{Precision for all calculations.}

\item{x_PE}{The range of used temperatures.}

\item{PE_genes}{Vector of 25 values that define the PE reaction norm.}

\item{GW_genes}{Vector of 5 vales that define the GW reaction norm.}

\item{OW_genes}{Vector of 5 vales that define the OW reaction norm.}

\item{PW_genes}{Vector of 5 vales that define the PW reaction norm.}

\item{mean_surv}{Genetic value for the survival midpoint.}

\item{fecundity_genes}{Vector of 3 values that define the mean,
sd and skew for fecundity.}
}
\description{
Constructor method for fly
}
