% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{get_fecundity}
\alias{get_fecundity}
\title{fly method to retrieve fecundity with temperature orr day}
\usage{
get_fecundity(object, temperature, day)
}
\arguments{
\item{object}{An object}

\item{temperature}{Value of temperature for which to retrieve fecundity.}

\item{day}{Day for which to retrieve fecundity.}
}
\description{
fly method to retrieve fecundity with temperature orr day
}
