% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly.R
\name{get_fecundity}
\alias{get_fecundity}
\alias{get_fecundity,fly-method}
\title{fly method to retrieve fecundity with day}
\usage{
get_fecundity(object, temperature)

\S4method{get_fecundity}{fly}(object, temperature)
}
\arguments{
\item{object}{An object}

\item{temperature}{Temperature for which to retrieve fecundity.}
}
\value{
Fecundity at the provided temperature.
}
\description{
fly method to retrieve fecundity with day
}
